% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching-package.R
\docType{package}
\name{branching-package}
\alias{branching}
\alias{branching-package}
\title{Phylogenies from many sources}
\description{
Phylogenies from many sources
}
\author{
Scott Chamberlain <myrmecocystus@gmail.com>
}
\keyword{package}

