% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tol.R
\name{tol}
\alias{tol}
\title{Query Treebase for phylogenetic trees}
\usage{
tol(taxa, context_name = NULL, do_approximate_matching = TRUE, ids = NULL,
  include_deprecated = FALSE, include_dubious = FALSE, parser = "rncl",
  ...)
}
\arguments{
\item{taxa}{A list or character vector of taxon names}

\item{context_name}{name of the taxonomic context to be searched (length-one character vector)}

\item{do_approximate_matching}{A boolean indicating whether or not to perform approximate string (a.k.a. "fuzzy") matching. Will greatly improve speed if this is turned OFF (false). By default, however, it is on (true).}

\item{ids}{An array of ids to use for identifying names. These will be assigned to each name in the names array. If ids is provided, then ids and names must be identical in length.}

\item{include_deprecated}{A boolean indicating whether or not to include deprecated taxa in the search.}

\item{include_dubious}{Whether to include so-called 'dubious' taxa–those which are not accepted by OTT.}

\item{...}{Further args passed on to \code{\link[rotl]{tol_induced_subtree}}}
}
\value{
One or more trees of class \code{phylo}
}
\description{
Query Treebase for phylogenetic trees
}
\examples{
\dontrun{
taxa <- c("Poa annua", "Phlox diffusa", "Helianthus annuus")
tree <- tol(taxa)
plot(tree, no.margin=TRUE)
}
}

