% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brranching-package.R
\docType{data}
\name{tpl}
\alias{tpl}
\title{Lookup-table for family, genus, and species names for ThePlantList gymnosperms}
\format{A data frame with 23,801 rows and 2 variables:
\describe{
  \item{family}{family name}
  \item{genus}{genus name}
}}
\source{
\url{http://www.theplantlist.org/}
}
\description{
These names are from \url{http://www.theplantlist.org/}, collected
on 2015-11-11, and are from version 1.1 of their data. This data is
used in the function \code{\link{phylomatic_names}}.
}
\keyword{data}

