% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brranching-package.R
\docType{package}
\name{brranching-package}
\alias{brranching-package}
\alias{brranching}
\title{brranching}
\description{
Phylogenies from many sources
}
\author{
Scott Chamberlain \href{mailto:myrmecocystus@gmail.com}{myrmecocystus@gmail.com}
}
\keyword{package}
