% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brranching-package.R
\docType{package}
\name{brranching-package}
\alias{brranching-package}
\alias{brranching}
\title{Phylogenies from many sources}
\description{
Phylogenies from many sources
}
\author{
Scott Chamberlain <myrmecocystus@gmail.com>
}
\keyword{package}
